fstman: add HAL version 1.0
Add version 1.0 of the fstman HAL.
This HAL is served by the fst-manager daemon,
and allows the WIFI framework to control the
FST session and support features such as:
- Dynamic enslave of both master interface (wlan0)
and rate upgrade interface (wigig0). This allows
auto-detection and switching to FST connection
without special UI and disable/enable WIFI.
- Support renaming of group interfaces, this is
needed because the WIFI framework can assign
different names to the client interface (such
as wlan0/wlan1).
- Infrastructure for supporting fst-manager for
SoftAP in parallel with STA fst-manager.
- Support FST and WIFI SON modes.

Change-Id: I51b5c5eda9bdc401e25e915b295efa4abf2d50f1